requirements-dev.txt
# Development dependencies for my-py-bdd-app
# Include core requirements first
-r requirements.txt

# Code coverage and testing tools
pytest-cov>=4.0.0
coverage[toml]>=7.0.0

# Code quality and formatting tools
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Additional development tools (optional)
# pre-commit>=3.0.0  # Git hooks for code quality
# mypy>=1.0.0        # Static type checking
# bandit>=1.7.0      # Security vulnerability scanner